DemoGafferExtension : added TaskAlgo implementation
To allow for easier tracking of external dependencies in a Gaffer
script, I added the "TaskAlgo" submodule to the "GafferDemoExtension"
package. It has multiple functions to retrieve upstream file
dependencies, based on the node values of the "fileName" plug.
Additional plug names can be specified in the designated plug name set,
should the need arise. The file names are returned as strings in a
Python Set. If needed, the user can also retrieve a set with
substituted variables for a given context.